Minimal install script.
This will install golang, rust and then pull in a bunch of tools from
projectdiscovery.io along with custom wordlists and SecLists.

This script clones the repo it exists in the idea being that you get the latest
changes even if you haven't done a git pull recently.

I plan to use snapshots on digitalocean to have an already setup VPS but it is also
nice to have a kind of live update feature.