Remove Rust installation.
None of the tools I currently use rely on Rust. I think was going to be
using a tool that used Rust but I decided against it and forgot to
remove the install command.

This should save a bit of time on initial setup as the only language
runtime that will need to be installed is golang.