Make install script leaner and replace cewl with mksub.
Changes.

Replaced cewl with mksub for subdomain permutation generation.
Use clang instead of gcc for C compilation.
Remove SecLists clone as it was drastically increasing the time to install.
Upgrade the version of go lang used.

I came across a new tool https://github.com/trickest/mksub which I wanted to integrate into this script.

When I was looking at the script I noticed a good few improvements and fixes, so I decided to bundle them in with this change.