Small refactor.
I noticed that on some distros because the script was invoked with sudo
it would result in the tooling being installed in /root as that is the
home directory of the invoking user.

There are also instances when fuff is doing content discovery and a host
returns 200 OK but a body with "error" or "not found".

This results in the output being clogged up with mostly useless results.

I am filtering them out for now.

I have also updated the README and changed the script to be invoked via
runuser.