Add yum package manager support.
One of the benefits of the Go runtime is that once you have it installed
you can install packages on your OS regardless of the distro once the
base package requirements are installed.

Checking for apt or yum makes this work across a greater set of Linux
distros for very little extra work/code.