Make random output directory relative to HOME directory.
The temp directory used to store per scan results was being created in CWD instead of HOME directory.

This was causing scan output to reach the terminal but output sent to the filesystem silently encountered
an error message and continued.

Considering the rest of the tooling in this scan relies on the previous I will make a note to add fail fast
mechanisms so errors do not silently propagate.