Remove colour output for httpx results.
The terminal escape codes to represent colour was breaking the awk search
string that I use.

I also noticed a bug in that if a url has the numeric 200 in it.

It will end up in the bucket of 200.txt HTTP response code results.

By dropping the colour output it allows me to match on [200] which is not valid in
a URL.