Fix arch detection and use PDTM for project discovery go binaries.
Originally I was piping the output of uname into cut so I could extract the arch of the machine we are running on.

This was initially working which now I am wondering how as the field I was extracting is the year the kernel was compiled.

I piped into cut as I didn't realise uname -m is actually portable across machines/libs.

I am now using uname -m so there should be no more problems with arch detection.

ProjectDiscovery released a tool called PDTM which is a package manager for their tooling.

Instead of pulling the go binaries in myself I am using that manager to do the setup of their tooling.

I still need to pull in go binaries via go install for unrelated tooling.